use uint64, fix data cache hit rate

+++ cache.h
@@ -23,19 +23,19 @@ class SetAssociativeCache {
     };
 
     std::string name;         // Cache name for reporting
-    size_t numSets;           // Number of sets in the cache
-    size_t numWays;           // Number of ways per set (associativity)
-    size_t accesses;          // Access counter
-    size_t hits;              // Hit counter
+    UINT64 numSets;           // Number of sets in the cache
+    UINT64 numWays;           // Number of ways per set (associativity)
+    UINT64 accesses;          // Access counter
+    UINT64 hits;              // Hit counter
     UINT64 globalLruCounter;  // Global counter for LRU policy
     std::vector<std::vector<CacheEntry>> sets;  // Cache storage [set][way]
 
     // Find the LRU entry in a set
-    size_t findLruWay(size_t setIndex) const {
-        size_t lruWay = 0;
+    UINT64 findLruWay(UINT64 setIndex) const {
+        UINT64 lruWay = 0;
         UINT64 minCounter = sets[setIndex][0].lruCounter;
 
-        for (size_t way = 0; way < numWays; way++) {
+        for (UINT64 way = 0; way < numWays; way++) {
             if (!sets[setIndex][way].valid) {
                 return way;  // Return first invalid entry
             }
@@ -48,20 +48,20 @@ class SetAssociativeCache {
     }
 
     // Update LRU status for an entry
-    void updateLru(size_t setIndex, size_t wayIndex) {
+    void updateLru(UINT64 setIndex, UINT64 wayIndex) {
         sets[setIndex][wayIndex].lruCounter = ++globalLruCounter;
     }
 
     // Hash function to map from tag to set index
-    virtual size_t getSetIndex(const TagType& tag) const = 0;
+    virtual UINT64 getSetIndex(const TagType& tag) const = 0;
 
    protected:
     virtual void handleEviction(const TagType& tag, const ValueType& value,
                                 bool dirty) = 0;  // Pure virtual function
 
    public:
-    SetAssociativeCache(const std::string& cacheName, size_t num_sets,
-                        size_t ways)
+    SetAssociativeCache(const std::string& cacheName, UINT64 num_sets,
+                        UINT64 ways)
         : name(cacheName),
           numSets(num_sets),
           numWays(ways),
@@ -71,9 +71,9 @@ class SetAssociativeCache {
 
         // Initialize cache structure
         sets.resize(numSets);
-        for (size_t i = 0; i < numSets; i++) {
+        for (UINT64 i = 0; i < numSets; i++) {
             sets[i].resize(numWays);
-            for (size_t j = 0; j < numWays; j++) {
+            for (UINT64 j = 0; j < numWays; j++) {
                 sets[i][j].valid = false;
             }
         }
@@ -84,9 +84,9 @@ class SetAssociativeCache {
     // Core lookup operation
     bool lookup(const TagType& tag, ValueType& value) {
         accesses++;
-        size_t setIndex = getSetIndex(tag);
+        UINT64 setIndex = getSetIndex(tag);
 
-        for (size_t way = 0; way < numWays; way++) {
+        for (UINT64 way = 0; way < numWays; way++) {
             if (sets[setIndex][way].valid && sets[setIndex][way].tag == tag) {
                 hits++;
                 value = sets[setIndex][way].value;
@@ -100,10 +100,10 @@ class SetAssociativeCache {
     // In cache.h, inside SetAssociativeCache class:
     void insert(const TagType& tag, const ValueType& value,
                 bool isWrite = false) {
-        size_t setIndex = getSetIndex(tag);
+        UINT64 setIndex = getSetIndex(tag);
 
         // If block already in cache, update value and mark dirty on write
-        for (size_t way = 0; way < numWays; ++way) {
+        for (UINT64 way = 0; way < numWays; ++way) {
             if (sets[setIndex][way].valid && sets[setIndex][way].tag == tag) {
                 sets[setIndex][way].value = value;
                 if (isWrite) {
@@ -116,7 +116,7 @@ class SetAssociativeCache {
         }
 
         // Block not present – choose a victim to evict (LRU)
-        size_t victimWay = findLruWay(setIndex);
+        UINT64 victimWay = findLruWay(setIndex);
         bool evictValid = sets[setIndex][victimWay].valid;
         bool evictDirty = false;
         TagType evictTag;
@@ -143,13 +143,13 @@ class SetAssociativeCache {
     }
 
     // Stats reporting methods
-    size_t getAccesses() const { return accesses; }
-    size_t getHits() const { return hits; }
+    UINT64 getAccesses() const { return accesses; }
+    UINT64 getHits() const { return hits; }
     double getHitRate() const {
         return accesses > 0 ? (double)hits / accesses : 0.0;
     }
     const std::string& getName() const { return name; }
-    size_t getSize() const { return numSets * numWays; }
-    size_t getNumSets() const { return numSets; }
-    size_t getNumWays() const { return numWays; }
+    UINT64 getSize() const { return numSets * numWays; }
+    UINT64 getNumSets() const { return numSets; }
+    UINT64 getNumWays() const { return numWays; }
 };
